cleanups regarding the merge
