+++ src/iblphotometry/outlier_detection.py
@@ -123,15 +123,6 @@ def _remove_spikes(F: nap.Tsd, sd: int = 5, w: int = 25):
     return nap.Tsd(t=t, d=y)
 
 
-# def remove_spikes_(fp_frame: nap.TsdFrame, sd: int = 5, w: int = 25):
-#     # FIXME REFACTOR very ugly multichannel version
-#     # possible solution: overloading, if Tsd, run once, if TsdFrame, operate on all columns
-#     f = np.zeros_like(fp_frame.values)
-#     for i, col in enumerate(fp_frame.columns):
-#         f[:, i] = remove_spikes(fp_frame[col], sd=sd, w=w).values
-#     return nap.TsdFrame(t=fp_frame.times(), d=f, columns=fp_frame.columns)
-
-
 def remove_spikes(F: nap.Tsd | nap.TsdFrame, sd: int = 5, w: int = 25):
     if isinstance(F, nap.TsdFrame):
         f = np.zeros_like(F.values)
